+++ main/3d/src/Vayo3dNodeAnimator.cpp
@@ -15,7 +15,7 @@ NodeAnimator::NodeAnimator(const wstring& name, SceneManager* oriSceneMgr)
 	, _oriSceneMgr(oriSceneMgr)
 	, _name(name)
 {
-	static unsigned short idx = 0;
+	static unsigned long long idx = 0;
 	if (0 == _name.compare(L""))
 	{
 		std::wstringstream ss;
