+++ plugins/RenderSystem_GL/src/GLTesselator.cpp
@@ -24,7 +24,7 @@ GLTesselator::GLTesselator(const wstring& name)
 	: Tesselator(name)
 	, _tess(NULL)
 {
-	static unsigned short idx = 0;
+	static unsigned long long idx = 0;
 	if (0 == _name.compare(L""))
 	{
 		std::wstringstream ss;
